fix: Fix GA tag location
